+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Bug fixes:
 * Fixed issue with ``Kernel#` `` when invalid UTF-8 given (#2118).
 * Fixed issue with `Method#to_proc` and special variable storage (#2156).
 * Add missing `offset` parameter for `FFI::Pointer#put_array_of_*` (#1525).
+* Fixed issue with different `Struct`s having the same hash values (#2214).
 
 Compatibility: